back/requirements.txt
aiojobs==0.3.0
aioredis==1.3.1
astroid==2.4.2
async-timeout==3.0.1
click==7.1.2
colorama==0.4.4
cx-Oracle==8.0.1
ecdsa==0.14.1
fastapi==0.62.0
fastapi-plugins==0.6.0
greenlet==1.0.0
h11==0.11.0
h2==4.0.0
hiredis==1.1.0
hpack==4.0.0
hyperframe==6.0.0
isort==5.7.0
lazy-object-proxy==1.4.3
ldap3==2.9
mccabe==0.6.1
multipart==0.2.3
passlib==1.7.4
priority==1.3.0
pyasn1==0.4.8
pydantic==1.7.3
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-jose==3.2.0
python-multipart==0.0.5
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.22
sqlalchemy-filters==0.12.0
starlette==0.13.6
tenacity==7.0.0
toml==0.10.2
uvicorn==0.13.3
values
wrapt==1.12.1
wsproto==1.0.0
XlsxWriter==1.3.7
